requirements.txt
oandapyV20
pandas
numpy
backtrader
python-dotenv
click